+++ HISTORY.md
@@ -83,6 +83,7 @@
 ### 2023-07-18
 * lists: fixed accessing negative indices in linked lists 
 * v9-rc17
+* blocks: fixed a multi-arg slot insertion / deletion glitch
 
 ### 2023-07-17
 * Getters and setter library: fixed "stage size" reporter